# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fresher Resume Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  username: string;
  email: string;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Resume {
  id: number;
  userId: number;
  resumeName: string;
  templateId: number;
  isPublic: boolean;
  /** @nullable */
  publicUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ResumeInput {
  resumeName: string;
  templateId: number;
}

export interface ResumeUpdate {
  resumeName?: string;
  templateId?: number;
  isPublic?: boolean;
}

export interface PersonalInfo {
  id: number;
  resumeId: number;
  fullName: string;
  email: string;
  phone: string;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  portfolio?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  photoUrl?: string | null;
}

export interface Objective {
  id: number;
  resumeId: number;
  summaryText: string;
}

export interface Education {
  id: number;
  resumeId: number;
  institution: string;
  degree: string;
  fieldOfStudy: string;
  graduationYear: number;
  cgpa: string;
}

export type SkillProficiencyLevel = typeof SkillProficiencyLevel[keyof typeof SkillProficiencyLevel];


export const SkillProficiencyLevel = {
  Beginner: 'Beginner',
  Intermediate: 'Intermediate',
  Advanced: 'Advanced',
  Expert: 'Expert',
} as const;

export interface Skill {
  id: number;
  resumeId: number;
  skillName: string;
  proficiencyLevel: SkillProficiencyLevel;
}

export interface Project {
  id: number;
  resumeId: number;
  projectTitle: string;
  description: string;
  technologies: string;
  /** @nullable */
  projectLink?: string | null;
  /** @nullable */
  role?: string | null;
}

export interface Experience {
  id: number;
  resumeId: number;
  company: string;
  position: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  isCurrent: boolean;
  responsibilities: string;
}

export interface Certification {
  id: number;
  resumeId: number;
  certName: string;
  /** @nullable */
  issuingOrg?: string | null;
  /** @nullable */
  dateIssued?: string | null;
  /** @nullable */
  description?: string | null;
}

export type LanguageProficiency = typeof LanguageProficiency[keyof typeof LanguageProficiency];


export const LanguageProficiency = {
  Native: 'Native',
  Fluent: 'Fluent',
  Intermediate: 'Intermediate',
  Basic: 'Basic',
} as const;

export interface Language {
  id: number;
  resumeId: number;
  languageName: string;
  proficiency: LanguageProficiency;
}

export interface ResumeDetail {
  resume: Resume;
  personalInfo?: PersonalInfo;
  objective?: Objective;
  education: Education[];
  skills: Skill[];
  projects: Project[];
  experience: Experience[];
  certifications: Certification[];
  languages: Language[];
}

export interface PersonalInfoInput {
  fullName: string;
  email: string;
  phone: string;
  linkedin?: string;
  portfolio?: string;
  address?: string;
  photoUrl?: string;
}

export interface ObjectiveInput {
  /** @maxLength 500 */
  summaryText: string;
}

export interface EducationInput {
  institution: string;
  degree: string;
  fieldOfStudy: string;
  graduationYear: number;
  cgpa: string;
}

export type SkillInputProficiencyLevel = typeof SkillInputProficiencyLevel[keyof typeof SkillInputProficiencyLevel];


export const SkillInputProficiencyLevel = {
  Beginner: 'Beginner',
  Intermediate: 'Intermediate',
  Advanced: 'Advanced',
  Expert: 'Expert',
} as const;

export interface SkillInput {
  skillName: string;
  proficiencyLevel: SkillInputProficiencyLevel;
}

export interface ProjectInput {
  projectTitle: string;
  description: string;
  technologies: string;
  projectLink?: string;
  role?: string;
}

export interface ExperienceInput {
  company: string;
  position: string;
  startDate: string;
  endDate?: string;
  isCurrent: boolean;
  responsibilities: string;
}

export interface CertificationInput {
  certName: string;
  issuingOrg?: string;
  dateIssued?: string;
  description?: string;
}

export type LanguageInputProficiency = typeof LanguageInputProficiency[keyof typeof LanguageInputProficiency];


export const LanguageInputProficiency = {
  Native: 'Native',
  Fluent: 'Fluent',
  Intermediate: 'Intermediate',
  Basic: 'Basic',
} as const;

export interface LanguageInput {
  languageName: string;
  proficiency: LanguageInputProficiency;
}

export interface Template {
  id: number;
  templateName: string;
  templateStyle: string;
  colorScheme: string;
  fontFamily: string;
  description: string;
}

export interface AtsScore {
  score: number;
  tips: string[];
  keywordsFound: string[];
  missingKeywords: string[];
}

export interface ExportResult {
  pdfBase64: string;
  filename: string;
}

export type DashboardStatsTemplateUsageItem = {
  templateName: string;
  count: number;
};

export interface DashboardStats {
  totalResumes: number;
  totalDownloads: number;
  recentResumes: Resume[];
  templateUsage: DashboardStatsTemplateUsageItem[];
}